requirements.txt
django
djangorestframework
mysqlclient
uwsgi
celery
redis
numpy
face-recognition
isort